t fp16

+++ src/plugins/intel_gpu/src/graph/moe_gemm.cpp
@@ -37,7 +37,7 @@ std::vector<layout> moe_gemm_inst::calc_output_layouts(moe_gemm_node const& /*no
         output_shape = { ov::Dimension(m), ov::Dimension(n) };
     }
     std::cout << "calc_output_layouts" << std::endl;
-    auto output_layout = layout{ output_shape, data_types::f32, format::bfyx };
+    auto output_layout = layout{ output_shape, data_types::f16, format::bfyx };
     std::cout << output_layout.to_short_string() << std::endl;
     return {output_layout};
 }
